Ensure kill future is resolved after KillInterruption
The future created for the kill transport task upon a KillInterruption
was not properly set in the case of another interruption, exception or
successful termination. In addition, we make sure that the calculation
states are set in accordance with the progression of the process state.
For example, the calculation state is not altered in the kill transport
task, but will only be set in `on_killed` of the JobProcess, to guarantee
that the calculation state is only set when the killing was successfull.