Miscellaneous fixes to `verdi` output redirection (#1894)
* docs/qb/usage: fix syntax highlight

* cmdline/utils/echo: allow specific err=True/False

In the *nix world, stderr is not only used to output error messages, but
also a means to output information to the user while stdout is
redirected to a file.

Example: you want to dump some text-based data from the database to a
file. The simplest way is by using output redirection::

    verdi data myplug dump > output

But at the same time you want to inform the user how many records were
dumped. The solution is to write the data to stdout while writing the
statistics to stderr. But for that we have to be able to tell the `echo`
functions where to write.

* cmdline/utils/echo: add is_output_redirected()

* orm/impl/sqla/node: fix sphinx cross-ref warning

* cmdline/commands/verdi: show help if called without command