Persist `paused` future of `Process` to maintain it after loading (#1902)
If a process is paused, indicated by the `_paused` attribute being
set to a future, it should remain paused when that process is recreated
from a saved checkpoint. To achieve this, we turn the `_paused` future
into a persistable future and make sure that it is persisted for the
`Process` class.