Pause `JobProcess` when transport task falls through exponential backoff
All transport tasks for the `JobProcess` are wrapped in the exponential
backoff retry coroutine utility, which when an exception occurs during
the transport task, will reschedule the task with an exponential backoff.
However, the backoff has a maximum number of retries, which when hit would
bubble up the exception and causing the process to except. With the new
pausing functionality in place, instead we can catch the
`TransportTaskException` and pause the process. The user then has the chance
to investigate the logs to determine the problem. If the problem was just
of a temporary nature, the user can then resume the process. If instead
the failure was of unrecoverable nature, the user can always decide to kill
the process.