convert to new-style division where necessary
the changes by the 2to3 fixer needed to be carefully reviewed, since it
converted all `/` to `//`, which is definitely not what we want

After adding `from __future__ import division`, the division operator `/`
always returns a float. To get the integer division, use now `//`.

Since most of the time, the float division was intended, the commit
consists mostly of adding the import statement.