replace all calls to `map()` with list comprehensions
most of the time they are easier to comprehend, in some cases this is
even an optimization since the intermediary list can be removed (for
example when inside `join` clauses which also takes generator
expressions).

for the `cod` and `icsd` there is also some refactoring of stringify
functions to use `.format()` instead (increases readability further).