make sure all uses of `zip()` are future-proof
The import `from six.moves import zip` resolves to `itertools.izip` on
Python 2 and the built-in `zip` on Python 3.

This can actually give a speedup on Python 2 since a lot of `for` loops
are now using iterators instead of full lists.

There is a nice idiom here. Instead of converting the iterator returned
by new-style `zip` to a list and then get the first element by index
`[0]`::

  list(zip(a, b))[0]

one can use the `next()` function on the iterator instead::

  next(zip(a,b))

Care has been taken to ensure that every `for` loop over a `zip`
iterator does not modify the contents of the variables for the `zip`
(which would have side-effects in Python 3, but worked before in Python
2).