enable the `filter` fixer after all
This makes sure that people are using the Python 3 style `filter` and
since we're using `six.moves` anyway we can also do it for this one.

Also fixes one usage of `filter` which was only discovered by running
`python-modernize` using Python 3.

Could be squashed with 'replace/check current uses of `filter()`'