common/hashing: take care of string vs bytes
Using `latin1` as the encoding since that was the default in Python 2
and the `b` literal prefix which is ignored by Python 2 but takes care
of proper concatenation in Python 3.

Files are opened in binary, which does not change anything in Python 2
but prevents Python 3 from automatically doing a UTF-8 decode, which
fails for binary data.