Added minimum scheduler poll interval
Now the computer has a property that can be set to regulate the interval
with which the scheduler command to get jobs can be called.  This acts
as a minimum and may in practice be longer if the transport minimum open
interval is longer.

With these changes all jobs in a single runner on the same computer also
have their requests to get the jobs list batched and therefore there
should be a significant decrease in the number of calls asking the
scheduler to list the currently running jobs.