Fix inline function build_kind_name of StructureData.set_pymatgen_structure
The inline function `build_kind_name` used in the `set_pymatgen_structure`
method of the `StructureData` class is tasked with constructing a kind name
given a pymatgen object with the species and their weights, in the case that
any of the sites in the pymatgen structure did not already have a kind name.
The original implementation was using list comprehension variables outside
of its scope, which is not allowed in python3 and most likely indicated a
broken implementation. The documentation also mentioned that in the case of
no spin, it would simply return the specie, but in fact it returned None in
that case and on top of that, the documentation then fails to consider cases
with more than one specie.

The solution here is to first check if the kind is both composite and is
tagged as having spin, which is not supported and so it raises.

For the spin case, the returned value is then simply the specie name with
a digit appended to indicate the spin channel, for composites None is
returned.