Allow an OverridableOption to be cloned while overriding some kwargs
The OverridableOption allows to predefine reusable click options for
the CLI. Calling the constructor will store the args and the kwargs
and by calling the instance, the click option will be generated from
those args and kwargs, while giving the chance to override the kwargs.

However, in many situations, one wants to slightly adapt a predefined
option for a set of sub commands, that largely share the settings of
that predefined option, yet all need to have one or more kwargs
changed. Since the instance is already created and the defaults can
not be changed, this is not possible.

The solution is to define the `clone` method, that will create a
copy of the instance and override the keyword arguments that were
set during construction with those that are passed in the method call.

This concept has now been applied to deduplicate the command line
options for `verdi code setup` and `verdi code duplicate` as well as
their analogues for the `verdi computer` command. Many of their options
are already defined in `aiida.cmdline.params.options` so we wanted to
reuse them, however, the `setup` and `duplicate` both required quite a
few additional kwargs in the call. Which would not have been a problem
if it were not for the fact that `duplicate` in fact needs one more
keyword, namely the `contextual_default`. By using `clone` to clone
the shared predefined options into `code` or `computer` specific cli
options, the old call functionality of the Overridable option can now
be used to pass the `contextual_default` for the options of the
`duplicate` command.