Ensure WorkChain does not exit unless stepper returns non-zero value
The engine exposes the possibility to the user of aborting a running
workchain from within its outline step functions, by simply returning
a non-zero integer or `ExitCode` instance with a non-zero status.
Any other return value should be ignored in terms of aborting the
running of the `WorkChain`. Due to a logical flaw in the `_do_step`
implementation of the `WorkChain` this heuristic was not being
respected and returning an `ExitCode` instance with zero status, would
cause the `WorkChain` to exit the control flow.