Add warning message for `verdi process` commands if daemon not running (#1943)
We add the `only_if_daemon_running` decorator to all `verdi process`
commands that rely on the daemon running. Note, however, that we override
the default behavior of emitting a critical error and exiting when the
daemon is not running, because although most likely these operations
will time out if the daemon is not running, it is not guaranteed. Even
if a user runs a process in a normal runner (i.e. not a runner launched
and controlled by the daemon) that process is still reachable through
the `verdi process` commands, since all runners have RabbitMQ comms.
However, since this usecase will probably be rare, and most of the time
users will only run processes through daemon runners, emitting a warning
when the daemon is not running, to indicate that the call will most
likely timeout because the process is not reachable, seems in order.

Note that these changes required a change in the `only_if_daemon_running`
decorator to allow the message and exit behavior to be customized.