Fix the help message generator of InteractiveOption for Choice params
The `InteractiveOption.format_help_message` method was expecting the list of
choices, in the case of a `Choice` parameter type, to be a list of tuples where
the first element is the name of the value and the second a description. However,
the default `complete` method for the `click.Choice`, which is defined by the
`click-completion` plugin, returns a flat list. Here we update the format method
to be able to deal with both a flat list as well as a list of tuples.