Fix the toggling of help message for InteractiveOptions (#1957)
The triggering of help messages for InteractiveOption cli params was broken.
The problem is that the prompt, which is defined by `click.prompt`, which is
used to obtain a value from the user, calls the conversion of the string value
from the command line to the intended type (defined by the type and default of
the parameter), before it is returned to the `simple_prompt_loop` of the custom
InteractiveOption. This means that for any option with a non-string type, attempts
by the user to trigger the display of the help message by entering the '?'
character will result in an exception, because the click prompt will try to convert
the question mark first, before the internal prompt gets the chance to check for
the presence of this predefined control character.

The `click.prompt` provides a hook to define the function used for the conversion
of the raw value called `value_proc`, which by default will be determined based
on the type and default of the option. As a solution we define a custom `value_proc`
function that will simply return the value, thereby skipping the conversion in
the click prompt loop, such that we can take care of it ourselves in our own
`simple_prompt_loop`.