Implement method for JobCalculation to create a restart builder
The `get_builder` method will return a "clean" builder for the given
`JobProcess`, without any inputs defined. Here we implement `get_builder_restart`
for the `JobCalculation` that will also return a `JobProcessBuilder` but with
the inputs and options pre-set with those that were used for the original node.
The builder can then immediately be resubmitted to run another instance of that
calculation, while of course giving the chance to change one or more inputs.

For example:

    calculation = load_node(<pk>)  # Some completed JobCalculation node
    builder = calculation.get_builder_restart()
    results = run(builder)