fix bug in load_dbenv() (#1967)
Fixes #1966
The if_dbenv_loaded() function was always returning True
if it is was called right after load_dbenv(). The problem was in
load_dbenv() function which was setting LOAD_DBENV_CALLED variable
to True before actually doing something. Consequently, for the
case where a connection to the database was not setup the call of
if_dbenv_loaded() was returning a wrong True result.