Improve error message if `load_dbenv` fails in click command
When the database environment is loaded for a click command through
the `with_dbenv` decorator and the database version is inconsistent
an exception is raised for the Django backend. Here we catch this
exception and instead print a critical error message. This has the
benefit that the user is not overwhelmed with the stack trace but
simply sees the instructions on how to perform the migration.