Ensure correct profile is printed if `check_schema_version` fails
The function was calling `get_current_profile` to determine the current
profile name, in order to format the migrate command that the user
should run in order to migrate the database. However, this function
only returns the profile name if the database environment was loaded,
and otherwise None. The problem is that the very reason we are trying
to get the current profile is *because* loading the database environment
failed. Therefore the returned profile will necessarily always be None.
Instead, we pass the profile explicitly from the `load_dbenv` call and
if it is None, the default profile was used, so we invoke the helper
function `get_default_profile` to get the applicable profile name.