Add `verdi database integrity` to the command line interface
The reason for this addition is that due to a bug in the database
models, due to the omission of a uniqueness condition on the UUID
column of the node table, nodes with duplicate UUIDs coud be created
leaving the database in an inconsistent state. Before we can apply
a migration to add the uniqueness constraint, any databases with
duplicate UUIDS would have to manually solve this or the migration
would fail.

A script was written for this particular case, but more such cases
could arise in the future. For this reason, this new `verdi` end
point was created to host any database integrity checks that might
be added in the future.