Add a node UUID uniqueness check before applying migration
A migration was recently added to apply a uniqueness constraint on
the UUID column of the Node table. However, for databases that violate
this constraint, the migration will fail and this will have to be
addressed first. In a previous commit a cli end point was added to
investigate the database for this problem and automatically fix it
to remove this integrity violation. Here we add a check before the
migration, to see if duplicate UUIDs exist. If that is the case, the
migration is interrupted and the user is instructed to run the
relevant verdi command to fix the problem first manually.