Add unit test for duplicate node UUID migration for Django
We define a special class `TestMigrations` that can be used to test
migrations for the Django backend. A test that implements this class
defines the starting revision and the migration it will apply and the
`setUpBeforeMigration` method, which allows to add data before the
migration is applied.

We apply this concept here where we migrate to the penultimate, as of
writing migration, insert nodes with duplicate UUIDs, apply the fix
that would be run manually by the user to fix this problem, before the
actual migration is applied. Afterwards, we verify that the deduplication
executed before the migration had the desired and expected effect.