Fix 'migration' message	for default profile
The message was	showing	the command to run to
migrate	the current profile. However,
the `get_current_profile()` function returns always
`None` when called at that point (before calling
`load_dbenv`. Replaced with the correct attribute
to inspect.
Also, activate the linter of the file, and added
an explicit reference to the most recent version of
pylint-django that supports py2, that solves *some*
(but not all unfortunately) of the wrong errors by pylint.