Fix version of `kiwipy==0.2.1` (#2014)
Both `kiwipy` and `plumpy`, which has the former as a dependency, are
currently under active development. We need `plumpy==0.10.6` but that
has only a lower bound on `kiwipy` requirement, of which a new version
`0.3.0` was just released. This is however incompatible with `plumpy`
0.10.6 but is installed nonetheless, causing all the tests too fail.
Therefore we are temporarily fixing the version of `kiwipy` explicitly.