Separate JobProcess submit task in folder upload and scheduler submit (#1946)
Originally, the submit TransportTask of a JobProcess included both the
uploading of the work directory on the remote machine as well as the
call to the scheduler to submit the job. These have now been separated
into two separate TransportTasks, `upload` and `submit`, in order to
provide finer grained control over the job process control flow.