Cleanup TransportRequest if the opening of transport excepts (#2021)
When the `transport.open()` call excepts in the `do_open` callback of a
`TransportRequest`, the request should be removed from the `TransportQueue`
because otherwise, if the request still hasl open references, the request
will persist and will continue to pass the same transport future. Since
this future has already excepted, the `request_transport` contexts will
keep excepting. Here we remove the excepted transport request as soon as
the transport open call excepts, forcing a new request and transport
future to be constructed.