Flag attribute as modified in ModelWrapper setter to properly flush
This bug was found bceause the `computer_configure` to set auth parameters
through the `verdi computer configure` command, which calls the `computer_configure`
command from the `aiida.control.computer` module, had no persistent effect
for the SqlAlchemy backend. The unit test that has been added could not
reproduce the problem, however, the added change to flag the attribute as
modified in the `__setattr__` method of the `ModelWrapper` has fixed the
problem that appeared when configuring a computer through the command line.