Revert invoke of configure after `verdi computer setup`
After a successful `verdi computer setup` the computer has to be
configured. At the end of the setup command, the required command that
has to be called to configure would be printed, but this was recently
changed into a confirmation prompt, followed by a direct invocation of
the configure command. However, due to a bug, the invoked configuration
command does not prompt the user properly. While we await this bug to
be fixed, we revert to simply printing the command that the user has
to call manually.