Use sub process to run daemon for `verdi process` tests
The unit tests for `verdi process` used to use separate threads to
run the daemon runner and the runner used to submit a test process,
however, using threads with SqlAlchemy is not supported, because they
use separate sessions and are typically not commensurate even though
the unit tests were assuming they are. Instead, we run a daemon runner
in a separate process, which simulates the real scenario as closely
as possible, while also avoiding the SqlAlchemy session problems.