Clarify the behavior of `verdi computer configure show` (#2056)
By default the command will now show the actually configured values
for a given computer and the defaults can be toggled through the
`--defaults` option. In addition, a bug is fixed when printing the
values of options that are not present in the auth info. This used
to assume they are always present which is not the case, resulting
in KeyError exceptions.