Prepare release v1.0.0a3 (#2061)
Updated the CHANGELOG.md and ran the header helper scripts.

Note that the helper scripts were updated to be able to cope with the new
`pylint` directives. The convention is that the header starts with an optional
shebang line, followed by the encoding declaration. Directly after, the copyright
message will be printed. Any module docstrings and linter declarations should
be placed *after* the copyright notice.