Added functionality to query for scheduler jobs by id
By default the `JobsList` will get the status of running jobs by requesting
the status for the user, associated with the auth info, however, some
schedulers do not support this functionality. Instead, one has to query for
a list of job ids. In this case the `JobsList` will simply get the job id list
from the internal mapping it keeps.