Log exception if DaemonRunner fails to instantiate in `start_daemon` (#2073)
If an exception occurs during instantiation of the `DaemonRunner` class
in `aiida.daemon.runner.start_daemon` the process will terminated. If
this was launched as a worker by the circus daemon, circus will simply
try to restart the worker, causing the exact same problem. Since the
exception is nog logged, the user is nonethewiser even though the
daemon appears to be running. However, it is continuously recreating
workers that instantly die.