Fix flake8 errors and warnings for python3
The command that was run is the following:

	flake8 . --count --select=E901,E999,F821,F822,F823 --show-source --statistics

All the warnings and errors have been resolved except for the usage of `long` in
the `aiida.backend.tests.nodes` file. This is already properly guarded by `six.PY2`
but flake8 clearly does not interpret this clause. We opt to ignore this as the
error is not applicable and we do not explicitly include these flake tests in the
pre-commit tests.