Implement the `verdi export inspect` command
This command enables a user to easily inspect the contents of an export
archive. The implementation uses a new class `Archive` which is a utility
class to deal with export archives. It should be used in a context manager:

    with Archive('/some/archive.aiida') as archive:
        version = archive.version_format

The context manager will ensure that the archive is lazily unpacked into a
sandbox folder which will be automatically cleaned when the context is left.

The command by default prints a summary of the contents and the version numbers
of the format itself and AiiDA with which it was exported. With flags one can
toggle to print just the format version, or the contents of the data or metadata
files contained within the archive.