Check for spurious output at bash login shell (#2132)
This is a different solution than PR #1880 and supercedes it.
We both document what the user shell should NOT do (produce
output if in interactive mode), and explain how to keep this but
guard it so that no output is produced at least in non-interactive
mode. Moreover, we add a test in `verdi computer test` that has
pointer to the documentation and the issue, making it (hopefully)
clear early on to the user on how to solve the issue.