Fix cyclic dependency in REST API translators
The CalculationTranslator is a sub class of the NodeTranslator, but the
latter was importing the former in the `get_retrieved_inputs` and the
`get_retrieved_ouputs` methods. These methods are called in the `_get_content`
method of the `NodeTranslator` but are only meaningful and only really
implemented for the `CalculationTranslator`. Since the latter is a sub class
simply overriding it will yield the correct result when the `_get_content`
method is called as it will be overridden.