Increase the length of partial UUID in ORM loader util tests (#2139)
With a partial UUID length of only two characters, the chance of having
multiple nodes in the test database with the same leading characters is
not negligible, leading to random failures of the test. The idea of the
test was to test the loading through partial UUIDs without a dash, which
for the current UUID version is the first eight characters. The chances
of having two UUIDs starting with the same eight characters should be
negligible.