Refactor the `Profile`, `ProfileConfig` and `DaemonClient` classes (#2144)
The `ProfileConfig` class is merged into the `Profile` class as the single
class to represent a profile. The config dictionary from the configuration
file is stored as the `config` property. The `Profile` class now also implements
the `is_test_profile` property, which should be the single point of truth when
determining whether a profile is a test profile. This is currently still based
on whether its name starts with `test_`, but by centralizing this question in
this class, it will be easy to change in the future when we device a more robust
definition of a test profile.

The `DaemonClient` now takes a `Profile` instance on construction instead of
sub classing it. To instantiate a `Profile` or `DaemonClient` instance, the
helper functions `get_profile` and `get_daemon_client` should be used, which
take an optional profile name and if left unspecified will use the current
default profile.