Refactor the creation of communicators and controllers
Just like a runner, the communicator for RabbitMQ as well as a
controller, that uses a communicator instance for RPCs to processes over
RabbitMQ, should probably have a global instance that is lazily
created. Parts of the code that require one of these entities most
likely will need the default one and we don't want it to be constructed
each time over.

This was necessary in order to be able to separate the responsability of
the instantiation of the process and sending the continuation task to
RabbitMQ in the `aiida.work.launch.submit` free function. The separation
allows us to use the default runner and therewith `get_runner` instead
of having to construct a new runner with a communicator. This in turn
solves the issue of many runners being constructed if `submit` is called
in a loop.