Raise `plumpy.PersistenceError` when saving or loading checkpoint excepts
Some of the calls in `save_checkpoint` and `load_checkpoint` in
`AiiDAPersister` can except. The runner or process that will call this
to either load or save a checkpoint will catch the `PersistenceError`
but any other exception might be ignored and disappear silently. Here we
explicitly catch these exceptions and raise `PersistenceError` instead.

We also set logging handlers for the `kiwipy` module with default level
set to `WARNING` such that they will be logged to the daemon logger.