Store any unstored nodes in workchain context before state transition
The user is allowed to place unstored nodes in the context. As a result,
the engine needs to ensure that before a workchain transitions from one
step to the other, those nodes are stored, because after the state
transition the process will be saved. In order to do so, a checkpoint
will be created of the serialized process instance. A checkpoint can
only be reloaded if all its nodes are stored in the database.