Set values for hard coded RabbitMQ and Runner parameters
The runner has a poll interval, which should not be 0 as that will lead
to the daemon runners using a lot of processor resources. The two
important settings for the RabbitMQ connection are the prefetch count
and the heartbeat timeout. The former is still experimental. Setting it
too high can lead to a daemon runner being overwhelmed, whereas setting
it too low can cause a the daemon to be deadlocked. The heartbeat
timeout is set to the maximum value that will be accepted anyway by a
default server configuration. The changes in kiwipy should make the
dropping of the heartbeat very unlikely anyway, but this is just an
extra precaution.

Note that all these parameters will have to become configurable on a per
profile basis at some point.