Created AiiDAManager to help with profile specific construction
We have a lot of places in aiida where we construct objects based on the
current profile settings.  A bunch of these have been moved to one
place, AiiDAManager, that acts as a global singleton that facilitates
the construction of objects based on the current profile (e.g. RMQ
communicators) as well as storing providing a place to put global
instance of objects that can be shared, e.g. the persister.