Wrap `flag_modified` to unwrap `ModelWrapper` instances
The new backend wraps the database model instances in the `ModelWrapper`
utility class, which will take care of timely flushing and updating the
model instance upon changes. Across the code, model instances are passed
to the `flag_modified` functionn of SqlAlchemy that mark a certain key
on an instance as modified. However, sometimes instead of the actual
model instance being passed, the wrapped version is passed. This used to
be ignored for `sqlalchemy==1.0.19` but in `sqlalchemy==1.2.12` to which
we have updated, an `InvalidRequestError` is raised.

To overcome this, we define our own `flag_modified` function that wraps
the one of SqlAlchemy and will unwrap any wrapped instances that are
passed to it.