Only let the `ModelWrapper` call `flag_modified`
The `_set_metadata` method of `SqlaComputer` was calling `flag_modified`
manually after having set the attribute, however, since the model is
wrapped in the `ModelWrapper`, it's `__setattr__` will take care of
calling `flag_modified` when it flushes upon an attribute being updated
for a stored entity. Calling `flag_modified` on the same key twice will
except in SqlAlchemy 1.2.12 because the "dirty" attribute has already
been removed and moved into the database model instance.