Set `rmq_submit=True` for the submit runner in `verdi process` tests
The default runner will have `rmq_submit=False` and so the test process
would be scheduled on the submission runner and would never get run.
Instead we configure the runner to submit over RabbitMQ which will cause
the process to be picked up by the daemon runner that is launched in a
sub process.