Allow the runner poll interval to be configurable per profile
This was necessary, because the default poll interval should be
non-zero, however, for testing we want it to be zero, for the tests to
be as fast as possible. The AiiDAManager will already ensure that
runners created for a test profile will have the poll interval set to
zero, but the daemon tests on Travis are performed with a normal profile
and so their daemon runners were getting the default non-zero values.

The only way to set the poll interval for the normal profile to a non
default is to make it a configurable property of the profile. In
addition to this change, in the `setup_profiles.sh` CI script, after
creating the profile, the poll interval is set to 0.