Compile the scripts before calling exec in `verdi run` (#2166)
This extra step is necessary for python modules like `inspect` to
function normally. Without this change, running a script with `verdi run`
that defines a workfunction and then runs it would fail, because inspect
would fail to determine the path of the source file of the workfunction,
i.e. the script that was passed to `verdi run`. This change fixes that.