Implement the new base ORM classes for process nodes
This commit defines the new basic hierarcy of the ORM classes for process nodes.
In addition, it already moves over the functionality of the current abstract
calculation classes to their new corresponding counterpart.

 - `AbstractCalculation` -> `ProcessNode` and `CalcJobNode`
 - `AbstractJobCalculation` -> `CalcJobNode`

This should allow the refactoring of the code according to the renaming of the
node ORM classes to begin.