Add temporary support for loading of new process node ORM classes
The loading of the correct ORM class for a given entry in the `DbNode` table
is based on the `type` column, but its interpretation is still in a half-way
state. The plugin loader will first attempt to reverse engineer it into an entry
point and if that works load the correspondingly class. If that fails, it is
interpreted as an internal class and will attempt to load it directly as a module
path.

Adding the new process node ORM classes, will need to be supported by the plugin
loader, so here we add their entry points. We do this in the `aiida:node` group
but note that in the future the entire system will be changed to solely rely on
entry points, at which the format of the `type` column can be changed to be a
fully qualified entry point string, just as the `process_type` column already is.