move to UUIDField of django 1.11
In the django backend, we were using the UUIDField of django-extensions
(the only use of django-extensions in the codebase).
Since django now provides a UUIDField, django-extensions discontinued
it and we have to move as well.

There are a few changes:

 * the old UUIDField had an "auto" keyword which resulted in a uuid
being automatically generated if none was provided at creation.
  In the new UUIDField, this should be the default behavior (?)
 * the old UUIDField had a "version" keyword which allowed to select the
UUID version.
  The new UUIDField instead has a "default" keyword, which allows to
provide a function to set the default value

For reference, this is the code of the old UUIDField in
django-extensions:
https://github.com/django-extensions/django-extensions/blob/cda632d6f5bd7ac7c63e2172789a55cd21dedd7a/django_extensions/db/fields/__init__.py#L371

And this is the new UUIDField in django 1.11:
https://github.com/django/django/blob/c7cc7526d5ee7d38a6ee1af03610f1aba1ea0c78/django/db/models/fields/__init__.py#L2291
    * this has no "auto" keyword (id=None should cause to use the 'default' when saving)
    * https://stackoverflow.com/a/35210896
        * the default is now "uuid.uuid4()"
        * this is the version we use in AiiDA, so we could just remove the "version" arg
(or we could provide our own UUIDField, where we hardcode version 4 (in case django changes their default in the future)
*