Drop the `DbCalcState` table (#2198)
The `DbCalcState` table was introduced to both keep the current job state
of `JobCalculations` but also to function as a sort of locking mechanism
to ensure only one process was operating on a job at a time, by enforcing
progressive state changes on the database level. In the new system, this
responsibility is taken care of by the `Runner` class that is running the
task corresponding to the job calculation. The calculation state is now
merely set as on the node for informationalpurposes to the user.

Therefore, the simpler solution is to set the job state as a simple attribute
and remove the `DbCalcState`. This also means that the `IMPORTED` calc state
is now obsolete, as that was needed to prevent imported calculations from
being picked up by the daemon, which used to determine what calculations to
run based on its calculation state.